add default user and fix bug

+++ src/main/java/lk/mydentist/api/controller/OrderController.java
@@ -2,7 +2,8 @@ package lk.mydentist.api.controller;
 
 import lk.mydentist.api.dto.OrderDto;
 import lk.mydentist.api.service.OrderService;
-import org.modelmapper.ModelMapper;
+import lk.mydentist.api.util.JWTTokenGenerator;
+import lk.mydentist.api.util.TokenStatus;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,34 +15,50 @@ import java.util.List;
 @RequestMapping("/order")
 public class OrderController {
     private final OrderService orderService;
-    private final ModelMapper modelMapper;
+    private final JWTTokenGenerator jwtTokenGenerator;
 
-    public OrderController(OrderService orderService, ModelMapper modelMapper) {
+    public OrderController(OrderService orderService, JWTTokenGenerator jwtTokenGenerator) {
         this.orderService = orderService;
-        this.modelMapper = modelMapper;
+        this.jwtTokenGenerator = jwtTokenGenerator;
     }
 
     @PostMapping
-    public ResponseEntity<Object> saveOrder(@RequestBody OrderDto orderDto){
-        OrderDto dto = this.orderService.saveOrder(orderDto);
-        return new ResponseEntity<>(dto, HttpStatus.CREATED);
+    public ResponseEntity<Object> saveOrder(@RequestBody OrderDto orderDto, @RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            OrderDto dto = this.orderService.saveOrder(orderDto);
+            return new ResponseEntity<>(dto, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @GetMapping
-    public ResponseEntity<Object> getAllOrders(){
-        List<OrderDto> dtos = this.orderService.getAllOrders();
-        return new ResponseEntity<>(dtos, HttpStatus.OK);
+    public ResponseEntity<Object> getAllOrders(@RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            List<OrderDto> allOrders = this.orderService.getAllOrders();
+            return new ResponseEntity<>(allOrders, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PutMapping("/{orderId}")
-    public  ResponseEntity<Object> updateOrder(@PathVariable Long orderId, @RequestBody OrderDto updateDto){
-        OrderDto dto = this.orderService.updateOrder(orderId,updateDto);
-        return new ResponseEntity<>(dto, HttpStatus.CREATED);
+    public ResponseEntity<Object> updateOrder(@PathVariable Long orderId, @RequestBody OrderDto updateDto, @RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            OrderDto dto = this.orderService.updateOrder(orderId, updateDto);
+            return new ResponseEntity<>(dto, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/{orderId}")
-    public ResponseEntity<Object> deleteOrder(@PathVariable Long orderId){
-        OrderDto dto = this.orderService.deleteOrder(orderId);
-        return new ResponseEntity<>(dto, HttpStatus.CREATED);
+    public ResponseEntity<Object> deleteOrder(@PathVariable Long orderId, @RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            OrderDto dto = this.orderService.deleteOrder(orderId);
+            return new ResponseEntity<>(dto, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 }
