+++ src/main/java/lk/mydentist/api/controller/RunnerController.java
@@ -3,6 +3,7 @@ package lk.mydentist.api.controller;
 import lk.mydentist.api.dto.RunnerDto;
 import lk.mydentist.api.service.RunnerService;
 import lk.mydentist.api.util.JWTTokenGenerator;
+import lk.mydentist.api.util.TokenStatus;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,28 +24,42 @@ public class RunnerController {
     }
 
     @PostMapping
-    public ResponseEntity<Object> saveRunne(@RequestBody RunnerDto runnerDto){
-
+    public ResponseEntity<Object> saveRunner(@RequestBody RunnerDto runnerDto, @RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
             RunnerDto dto = this.runnerService.saveRunner(runnerDto);
             return new ResponseEntity<>(dto, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @GetMapping
-    public ResponseEntity<Object> getAllRunners(){
-        List<RunnerDto> dtos = this.runnerService.getAllRunners();
-        return new ResponseEntity<>(dtos, HttpStatus.OK);
+    public ResponseEntity<Object> getAllRunners(@RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            List<RunnerDto> allRunners = this.runnerService.getAllRunners();
+            return new ResponseEntity<>(allRunners, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PutMapping("/{runnerId}")
-    public  ResponseEntity<Object> updateStudent(@PathVariable Long runnerId, @RequestBody RunnerDto updateRunnerDto){
-        RunnerDto dto = this.runnerService.updateRunner(runnerId,updateRunnerDto);
-        return new ResponseEntity<>(dto, HttpStatus.CREATED);
+    public ResponseEntity<Object> updateStudent(@PathVariable Long runnerId, @RequestBody RunnerDto updateRunnerDto, @RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            RunnerDto dto = this.runnerService.updateRunner(runnerId, updateRunnerDto);
+            return new ResponseEntity<>(dto, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/{runnerId}")
-    public ResponseEntity<Object> deleteStudent(@PathVariable Long runnerId){
-        RunnerDto dto = this.runnerService.deleteRunner(runnerId);
-        return new ResponseEntity<>(dto, HttpStatus.OK);
+    public ResponseEntity<Object> deleteStudent(@PathVariable Long runnerId, @RequestHeader(name = "Authorization") String authorizationHeader) {
+        if (this.jwtTokenGenerator.validateToken(authorizationHeader)) {
+            RunnerDto dto = this.runnerService.deleteRunner(runnerId);
+            return new ResponseEntity<>(dto, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(TokenStatus.TOKEN_INVALID, HttpStatus.UNAUTHORIZED);
+        }
     }
-
 }
